Add matching function for triangle similarity
The matching function takes the triangles found by the FindFeatures function
and finds correspondences based on the relative lengths of the edges of the
triangles.

One possible problem with this approach is that outliers (similar triangles in
completely different regions of the exposures) are not discarded. It would be
much better to take into account that the stars should move mostly uniformely.